pyproject.toml
[tool.poetry]
name = "testtask"
version = "0.1.0"
description = ""
authors = ["Vadym <umpire333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlmodel = "^0.0.16"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
